+++ chia/_tests/core/mempool/test_mempool_performance.py
@@ -2,9 +2,10 @@ from __future__ import annotations
 
 import pytest
 
-from chia._tests.util.misc import BenchmarkRunner, add_blocks_in_batches, wallet_height_at_least
+from chia._tests.util.misc import BenchmarkRunner, wallet_height_at_least
 from chia._tests.util.setup_nodes import OldSimulatorsAndWallets
 from chia._tests.util.time_out_assert import time_out_assert
+from chia.simulator.add_blocks_in_batches import add_blocks_in_batches
 from chia.types.full_block import FullBlock
 from chia.types.mempool_inclusion_status import MempoolInclusionStatus
 from chia.types.peer_info import PeerInfo
